+++ tests/Integration/CacheControl/CacheControlDirectiveTest.php
@@ -400,7 +400,7 @@ final class CacheControlDirectiveTest extends DBTestCase
                     }
                 }
             ', 'max-age=10, private',
-            ];
+        ];
 
         yield [/** @lang GraphQL */ '
                 {
@@ -412,6 +412,6 @@ final class CacheControlDirectiveTest extends DBTestCase
                     }
                 }
             ', 'max-age=20, private',
-            ];
+        ];
     }
 }