+++ CHANGELOG.md
@@ -9,6 +9,8 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+## v6.20.0
+
 ### Added
 
 - Allow `@cacheControl` directive on type level https://github.com/nuwave/lighthouse/pull/2451